# bpo-37207: Use PEP 590 vectorcall to speed up set()