--- a/Misc/NEWS.d/next/Core and Builtins/2020-03-15-23-16-00.bpo-37207.6XbnQA.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2020-03-15-23-16-00.bpo-37207.6XbnQA.rst
@@ -1,0 +1,2 @@
+Speed up calls to ``set()`` by using the :pep:`590` ``vectorcall`` calling
+convention. Patch by Dong-hee Na.