@vstinner 2020-03-16
IMHO it's worth it to add a few lines of C code to make tuple() and set() constructors faster.