@vstinner 2020-03-16
> Yeah, I think the PEP 590 implemented very well to bring performance enhancement very easily.

What I mean is that vectorcall should not be used for *everything*. But set() and tuple() are very commonly used, so here it's worth it.